requirements.txt

numpy
nltk
Pillow
